requirements.txt
# Flask и связанные библиотеки
Flask==3.1.1
Flask-CORS==6.0.1

# База данных
SQLAlchemy==2.0.41
PyMySQL==1.1.1

# Миграции
alembic==1.16.2

# Переменные окружения
python-dotenv==1.1.1

# Для разработки
Werkzeug==3.1.3 